Enable persistent queue in the build by default
