+++ CHANGELOG.md
@@ -38,6 +38,7 @@
 
 ### 💡 Enhancements 💡
 
+- Enable persistent queue in the build by default (#5828)
 - Bump to opentelemetry-proto v0.19.0. (#5823)
 - Expose `Scope.Attributes` in pdata (#5826)
