+++ exporter/exporterhelper/README.md
@@ -28,16 +28,14 @@ The following configuration options can be modified:
 
 ### Persistent Queue
 
-**Status: under development**
+**Status: [alpha]**
 
-> :warning: The capability is under development and currently can be enabled only in OpenTelemetry
-> Collector Contrib with `enable_unstable` build tag set. 
+> :warning: The capability is under development. To use it, a storage extension needs to be set up.
 
-With this build tag set, additional configuration option can be enabled:
+To use the persistent queue, the following setting needs to be set:
 
 - `sending_queue`
   - `storage` (default = none): When set, enables persistence and uses the component specified as a storage extension for the persistent queue
-    (note, `enable_unstable` build tag needs to be enabled first, see below for more details)
 
 The maximum number of batches stored to disk can be controlled using `sending_queue.queue_size` parameter (which,
 similarly as for in-memory buffering, defaults to 5000 batches).
@@ -111,3 +109,4 @@ service:
 ```
 
 [filestorage]: https://github.com/open-telemetry/opentelemetry-collector-contrib/tree/main/extension/storage/filestorage
+[alpha]: https://github.com/open-telemetry/opentelemetry-collector#alpha
