+++ exporter/exporterhelper/queued_retry_experimental_test.go
@@ -12,9 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//go:build enable_unstable
-// +build enable_unstable
-
 package exporterhelper
 
 import (
